Create a Debian package for this service
This package also installs a cronjob to update the GTFS database from
NTA once per day. It runs the script on setup to initialise the data
directory.

As part of this change, we now default to --use_python_zip. This gives us a
single file to ship (main.zip), and works around an issue in pkg_tar not
correctly including the Python runfiles.