Support services that cross day boundaries
GTFS allows for stop_times>2400hrs, which indicates the relevant time on
the following day. Instead of anchoring GetScheduledFor on the start date,
which was simple but always fragile, we look at a range of dates from
[start.date()-1 day, end.date()+1 day]. For each day, we compute the
possible service time, then filter based on the range and whether or not
it's a calendared day for the route.